Add files via upload
Initial files for SavyStudent project. makes sure to put the images in a folder titled "images" to display images. Important pages for back-end communication are dashboard, register, sign-on, update-budget, update-form, and update-income. Keep in mind that this is the initial draft and any issues you may find will be addressed at a later point. The primary focus for now was creating the pages needed to begin back-end communication.